requirements_qrauts.txt
streamlit
pandas
numpy
matplotlib
xlsxwriter
pillow
